First attempt to compute score for speculative candidates
Small change during the Scope Detection step which invoke the
Region Speculation. Here a decision is made if speculations should be
enabled for some candidate. To do so a score is computed which consists
of several parts (loops, conditionals, violations). If the command line
flag --spolly-dump is set the region together with the computed score
are written to a file.